Almamu's fix for non-integer skill rank

+++ src/eve-common/utils/EvEMath.cpp
@@ -10,7 +10,7 @@
 #include "../eve-common.h"
 
 // skill Equations
-uint32 EvEMath::Skill::PointsAtLevel(uint8 level, uint8 rank)
+uint32 EvEMath::Skill::PointsAtLevel(uint8 level, float rank)
 {
     if (level > EvESkill::MAXSKILLLEVEL)
         level = EvESkill::MAXSKILLLEVEL;
